requirements.txt
streamlit
pillow
numpy
pandas
opencv-python-headless
tensorflow==2.12.0